Match gd (thx Taka)

+++ src/dolphin/gd/GDPixel.c
@@ -7,13 +7,13 @@ void GDSetBlendMode(GXBlendMode type, GXBlendFactor src_factor,
 	GDWriteBPCmd(0xfe001fe3);
 	// clang-format off
 	GDWriteBPCmd(
-		((type == GX_BM_BLEND) || (type == GX_BM_SUBTRACT))
-		| (type == GX_BM_LOGIC) << 1
-		| dst_factor << 5
-		| src_factor << 8
-		| (type == GX_BM_SUBTRACT) << 11
-		| logic_op << 12
-		| 0x41 << 24
+		((u32)(type == GX_BM_BLEND || type == GX_BM_SUBTRACT)) |
+		((u32)(type == GX_BM_LOGIC)) << 1 |
+		((u32)dst_factor) << 5 |
+		((u32)src_factor) << 8 |
+		((u32)(type == GX_BM_SUBTRACT)) << 11 |
+		((u32)logic_op) << 12 |
+		((u32)0x41) << 24
 	);
 	// clang-format on
 }
@@ -25,16 +25,16 @@ void GDSetBlendModeEtc(GXBlendMode type, GXBlendFactor src_factor,
 {
 	// clang-format off
 	GDWriteBPCmd(
-		((type == GX_BM_BLEND) || (type == GX_BM_SUBTRACT))
-		| (type == GX_BM_LOGIC) << 1
-		| dither_enable << 2
-		| color_update_enable << 3
-		| alpha_update_enable << 4
-		| dst_factor << 5
-		| src_factor << 8
-		| (type == GX_BM_SUBTRACT) << 11
-		| logic_op << 12
-		| 0x41 << 24
+		((u32)(type == GX_BM_BLEND || type == GX_BM_SUBTRACT)) |
+		((u32)(type == GX_BM_LOGIC)) << 1 |
+		((u32)dither_enable) << 2 |
+		((u32)color_update_enable) << 3 |
+		((u32)alpha_update_enable) << 4 |
+		((u32)dst_factor) << 5 |
+		((u32)src_factor) << 8 |
+		((u32)(type == GX_BM_SUBTRACT)) << 11 |
+		((u32)logic_op) << 12 |
+		((u32)0x41) << 24
 	);
 	// clang-format on
 }
@@ -43,10 +43,10 @@ void GDSetZMode(u8 compare_enable, GXCompare func, u8 update_enable)
 {
 	// clang-format off
 	GDWriteBPCmd(
-		compare_enable
-		| func << 1
-		| update_enable << 4
-		| 1u << 30
+		((u32)compare_enable) |
+		((u32)func) << 1 |
+		((u32)update_enable) << 4 |
+		((u32)1) << 30
 	);
 	// clang-format on
 }
@@ -55,10 +55,10 @@ void GDSetDstAlpha(u8 enable, u8 alpha)
 {
 	// clang-format off
 	GDWriteBPCmd(
-		alpha
-		| enable << 8
-		| 1 << 25
-		| 1 << 30
+		((u32)alpha) |
+		((u32)enable) << 8 |
+		((u32)1) << 25 |
+		((u32)1) << 30
 	);
 	// clang-format on
 }
